+++ dcmrt/include/dcmtk/dcmrt/seq/drtcs.h
@@ -1,13 +1,13 @@
 /*
  *
  *  Copyright (C) 2008-2012, OFFIS e.V. and ICSMED AG, Oldenburg, Germany
- *  Copyright (C) 2013-2017, J. Riesmeier, Oldenburg, Germany
+ *  Copyright (C) 2013-2014, J. Riesmeier, Oldenburg, Germany
  *  All rights reserved.  See COPYRIGHT file for details.
  *
  *  Header file for class DRTContourSequence
  *
- *  Generated automatically from DICOM PS 3.3-2017e
- *  File created on 2017-12-05 09:30:54
+ *  Generated automatically from DICOM PS 3.3-2007
+ *  File created on 2014-03-15 16:58:36
  *
  */
 
@@ -56,7 +56,6 @@ class DCMTK_DCMRT_EXPORT DRTContourSequence
 
         /** assigment operator
          *  @param copy item object to be copied
-         *  @return reference to this object
          */
         Item &operator=(const Item &copy);
 
@@ -267,9 +266,9 @@ class DCMTK_DCMRT_EXPORT DRTContourSequence
 
         /// AttachedContours (3006,0049) vr=IS, vm=1-n, type=3
         DcmIntegerString AttachedContours;
-        /// ContourData (3006,0050) vr=DS, vm=3-3n, type=1
+        /// ContourData (3006,0050) vr=DS, vm=3-3n, type=1C
         DcmDecimalString ContourData;
-        /// ContourGeometricType (3006,0042) vr=CS, vm=1, type=1
+        /// ContourGeometricType (3006,0042) vr=CS, vm=1, type=1C
         DcmCodeString ContourGeometricType;
         /// ContourImageSequence (3006,0016) vr=SQ, vm=1, type=3
         DRTContourImageSequence ContourImageSequence;
@@ -279,7 +278,7 @@ class DCMTK_DCMRT_EXPORT DRTContourSequence
         DcmDecimalString ContourOffsetVector;
         /// ContourSlabThickness (3006,0044) vr=DS, vm=1, type=3
         DcmDecimalString ContourSlabThickness;
-        /// NumberOfContourPoints (3006,0046) vr=IS, vm=1, type=1
+        /// NumberOfContourPoints (3006,0046) vr=IS, vm=1, type=1C
         DcmIntegerString NumberOfContourPoints;
 
     };
@@ -302,7 +301,6 @@ class DCMTK_DCMRT_EXPORT DRTContourSequence
 
     /** assigment operator
      *  @param copy sequence object to be copied
-     *  @return reference to this object
      */
     DRTContourSequence &operator=(const DRTContourSequence &copy);
 
@@ -325,7 +323,7 @@ class DCMTK_DCMRT_EXPORT DRTContourSequence
     /** get number of items in the sequence
      *  @return number of items
      */
-    size_t getNumberOfItems() const;
+    unsigned long getNumberOfItems() const;
 
     /** goto first item in the sequence
      *  @return status, EC_Normal if successful, an error code otherwise
@@ -341,7 +339,7 @@ class DCMTK_DCMRT_EXPORT DRTContourSequence
      *  @param  num  number of the item to be selected (0..num-1)
      *  @return status, EC_Normal if successful, an error code otherwise
      */
-    OFCondition gotoItem(const size_t num);
+    OFCondition gotoItem(const unsigned long num);
 
     /** get current item in the sequence
      *  @param  item  reference to item pointer (result variable)
@@ -364,31 +362,31 @@ class DCMTK_DCMRT_EXPORT DRTContourSequence
      *  @param  item  reference to item pointer (result variable)
      *  @return status, EC_Normal if successful, an error code otherwise
      */
-    OFCondition getItem(const size_t num, Item *&item);
+    OFCondition getItem(const unsigned long num, Item *&item);
 
     /** get particular item in the sequence
      *  @param  num  number of the item to be retrieved (0..num-1)
      *  @return reference to specified item if successful, empty default item otherwise
      */
-    Item &getItem(const size_t num);
+    Item &getItem(const unsigned long num);
 
     /** get particular item in the sequence
      *  @param  num  number of the item to be retrieved (0..num-1)
      *  @return const reference to specified item if successful, empty default item otherwise
      */
-    const Item &getItem(const size_t num) const;
+    const Item &getItem(const unsigned long num) const;
 
     /** get particular item in the sequence
      *  @param  num  number of the item to be retrieved (0..num-1)
      *  @return reference to specified item if successful, empty default item otherwise
      */
-    Item &operator[](const size_t num);
+    Item &operator[](const unsigned long num);
 
     /** get particular item in the sequence
      *  @param  num  number of the item to be retrieved (0..num-1)
      *  @return const reference to specified item if successful, empty default item otherwise
      */
-    const Item &operator[](const size_t num) const;
+    const Item &operator[](const unsigned long num) const;
 
     /** add new item to the end of this sequence
      *  @param  item  reference to new item pointer (result variable)
@@ -401,13 +399,13 @@ class DCMTK_DCMRT_EXPORT DRTContourSequence
      *  @param  item  reference to new item pointer (result variable)
      *  @return status, EC_Normal if successful, an error code otherwise
      */
-    OFCondition insertItem(const size_t pos, Item *&item);
+    OFCondition insertItem(const unsigned long pos, Item *&item);
 
     /** remove particular item from the sequence
      *  @param  pos  position of the item to be removed (0..num-1)
      *  @return status, EC_Normal if successful, an error code otherwise
      */
-    OFCondition removeItem(const size_t pos);
+    OFCondition removeItem(const unsigned long pos);
 
   // --- input/output methods ---
 
@@ -442,7 +440,7 @@ class DCMTK_DCMRT_EXPORT DRTContourSequence
      *  @param  iterator  list iterator storing the position of the item
      *  @return status, EC_Normal if successful, an error code otherwise
      */
-    OFCondition gotoItem(const size_t num,
+    OFCondition gotoItem(const unsigned long num,
                          OFListIterator(Item *) &iterator);
 
     /** goto particular item in the sequence
@@ -450,7 +448,7 @@ class DCMTK_DCMRT_EXPORT DRTContourSequence
      *  @param  iterator  list iterator storing the position of the item
      *  @return status, EC_Normal if successful, an error code otherwise
      */
-    OFCondition gotoItem(const size_t num,
+    OFCondition gotoItem(const unsigned long num,
                          OFListConstIterator(Item *) &iterator) const;
 
   private:
